Fix issue with WBMP writer and image resizing.

+++ source/cpp/iohandler_wbmp.cpp
@@ -62,7 +62,7 @@ int CTL_WBMPIOHandler::WriteBitmap(LPCTSTR szFile, bool /*bOpenFile*/, int /*fhF
     {
         height = 255;
         width = 255;
-        m_pDib->ResampleDib(static_cast<double>(width), static_cast<double>(height));
+        m_pDib->ResampleDib(static_cast<long>(width), static_cast<long>(height));
         hDib = m_pDib->GetHandle();
     }
     else
